+++ q1simulator/trigger_sorting.py
@@ -9,8 +9,10 @@ logger = logging.getLogger(__name__)
 class SequencerTriggerInfo:
     module: int
     seq_number: int
-    trigger_addr: int
-    ''' address value of emitted trigger:  '''
+    emitted_triggers: int
+    ''' bits for emitted trigger.
+    0b00001010 for triggers 2 and 4.
+    '''
     used_triggers: int
     ''' bits for all triggers used by sequencer conditions.
     0b00001010 for triggers 2 and 4.
@@ -18,42 +20,54 @@ class SequencerTriggerInfo:
 
 def get_seq_trigger_info(module, seq_number, sequencer):
     if hasattr(sequencer,'thresholded_acq_trigger_en') and sequencer.thresholded_acq_trigger_en():
-        trigger_addr = sequencer.thresholded_acq_trigger_address()
+        emitted_triggers = 1 << (sequencer.thresholded_acq_trigger_address() - 1)
     else:
-        trigger_addr = 0
+        emitted_triggers = 0
     return SequencerTriggerInfo(
-            module, seq_number,
-            trigger_addr,
+            module,
+            seq_number,
+            emitted_triggers,
             sequencer.get_used_triggers()
             )
 
-def sort_sequencers(sequencers: List[SequencerTriggerInfo]):
-    seq_in = sequencers.copy()
-    seq_out = []
-
-    # check if there are not 2 sequencers with same emitted trigger.
-    emitters = defaultdict(int)
-    for seq in seq_in:
-        if seq.trigger_addr:
-            emitters[seq.trigger_addr] += 1
-            if emitters[seq.trigger_addr] > 1:
-                logger.warning(f'Multiple sequencers emit trigger {seq.trigger_addr}')
-
+def get_emitted_triggers(sequencers: List[SequencerTriggerInfo]):
     emitted = 0
-    while len(seq_in):
-        added = 0
-        for seq in seq_in.copy():
-            seq_requires = seq.used_triggers
-            seq_emits = 1 << (seq.trigger_addr - 1) if seq.trigger_addr else 0
-            if seq_emits and emitters[seq.trigger_addr] == 1:
-                # It's the only emitter, so it can use it's own output.
-                seq_requires &= ~seq_emits
-            if (seq_requires & ~emitted) == 0:
-                emitted |= seq_emits
-                seq_out.append(seq)
-                seq_in.remove(seq)
-                added += 1
-        if added == 0:
-            raise Exception(f"sorting failed... {seq_in}")
+    for seq in sequencers:
+        emitted |= seq.emitted_triggers
+    return emitted
 
-    return seq_out
+def sort_sequencers(sequencers: List[SequencerTriggerInfo]):
+    ''' Sorts sequencers putting trigger producers before trigger consumers.
+
+    Sequencers are executed sequentially. A sequencer must be executed after
+    all sequencers that emit triggers that it uses.
+
+    Notes:
+        * A sequencer may use a trigger not emitted by any sequencer.
+        * A sequencer may use a trigger that it also emits.
+    '''
+    result = []
+    unsorted = sequencers.copy()
+    while len(unsorted) > 0:
+        n_start = len(unsorted)
+        to_be_emitted = get_emitted_triggers(unsorted)
+        for seq in unsorted.copy():
+            if seq.used_triggers & to_be_emitted == 0:
+                # sequence does not use triggers that will be emitted after it.
+                result.append(seq)
+                unsorted.remove(seq)
+            elif seq.used_triggers & to_be_emitted & ~seq.emitted_triggers == 0:
+                # sequence emits a triggers it uses. Check if there are other sequencers emitting the trigger.
+                remaining = unsorted.copy()
+                remaining.remove(seq)
+                rem_emitted = get_emitted_triggers(remaining)
+                if seq.used_triggers & rem_emitted == 0:
+                    # OK. It does not use triggers of remaining sequencers.
+                    result.append(seq)
+                    unsorted.remove(seq)
+
+        if len(unsorted) == n_start:
+            raise Exception(f"Sorting sequencers on triggers failed. Unsorted: {unsorted}")
+
+    print('Sorted', result)
+    return result